refactor(uikit): move setupUpdateHasFocus into a11y/focus.ts
Hoists the focus-mirror primitive out of the Input-specific hidden-input module so every
component's hidden a11y element can drive the base-class hasFocus signal (Phase 0, T0.2).
hidden-input.ts re-exports it — no behavior change, no importer churn.
